[@ware]Modify ChangeToekn type from int to long
Former-commit-id: e85054480ac4052a353c5c7c50cf67a1a6732fcd

+++ nemakiware/src/main/java/jp/aegif/nemaki/service/dao/ContentDaoService.java
@@ -479,7 +479,7 @@ public interface ContentDaoService {
 	 *            "<= 0" means "infinite"
 	 * @return Return results with descending order by time
 	 */
-	List<Change> getLatestChanges(int startToken, int maxItems);
+	List<Change> getLatestChanges(long startToken, int maxItems);
 
 	/**
 	 * Create a change event
